requirements.txt
openai>=1.0.0
python-docx>=0.8.11
tenacity>=8.0.0
python-dotenv>=1.0.0
